pyproject.toml
[project]
name = "ai-agent-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "mcp[cli]>=1.27.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
